derive QueryId when generating mapping
