+++ tests_with_diesel_cli/src/with_default_schema.rs
@@ -24,26 +24,3 @@ pub fn insert(conn: &mut Conn, value: &Simple) -> Result<Simple, Error> {
         .values(value)
         .get_result(conn)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn round_trip() {
-        let mut conn = crate::get_connection();
-        let this = Simple {
-            id: 1,
-            some_value: MyEnum::Foo,
-        };
-        let that = insert(&mut conn, &this).unwrap();
-        assert_eq!(this, that);
-
-        // make a query that requires QueryId trait to exist
-        let _: Vec<Simple> = simple::table
-            .filter(simple::some_value.eq(MyEnum::Foo))
-            .limit(1)
-            .load(&mut conn)
-            .unwrap();
-    }
-}
